Renamed test_connection_client
